BJS2-54357: Рефакторинг
