+++ src/main/java/school/faang/user_service/publisher/AbstractEventPublisher.java
@@ -1,0 +1,14 @@
+package school.faang.user_service.publisher;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.core.RedisTemplate;
+
+@RequiredArgsConstructor
+public abstract class AbstractEventPublisher<T> {
+    private final RedisTemplate<String, Object> redisTemplate;
+    private final String channel;
+
+    public void publisherEvent(T event) {
+        redisTemplate.convertAndSend(channel, event);
+    }
+}
